Reformatted code with cogs
